had to revert uint8_t didn't work right

+++ ncrypt.c
@@ -23,8 +23,8 @@ usage:
 #define RB_LEFT 1
 #define RB_RIGHT 0
 
-uint8_t rotatebits(const uint8_t inchar, const int offset, const int lorr) {
-    uint8_t tmpchar;
+int rotatebits(const uint8_t inchar, const int offset, const int lorr) {
+    int tmpchar;
 
     switch (lorr) {
         case RB_LEFT:  tmpchar = ((inchar << offset) | (inchar >> (CHAR_BIT - offset))); break;
@@ -37,14 +37,14 @@ uint8_t rotatebits(const uint8_t inchar, const int offset, const int lorr) {
 
 int flipcrypt(FILE *fpin, FILE *fpout, const char *key, const int eord) {
 
-    uint8_t tmpchar;
+    int tmpchar;
     int keylen = strlen(key);
     int ik, broffset;
 
     ik = 0;
 
     tmpchar = fgetc(fpin);
-    while (tmpchar != (uint8_t)EOF) {
+    while (tmpchar != EOF) {
         
         // broffset determines bit rotation offset based on key value
         broffset = (key[ik] % CHAR_BIT);
